Revert "Try publishing in cjs format"
This reverts commit 7fbef817bb7be4334595f4ecdc30ba1f8764dbac. Without
type: module, we can't write an eslint that shows up in the editor. But
with it, we can't use magic strings in stylelint to import the config --
stylelint doesn't recognize mjs files. In the end, a slightly different
way of importing and using the rules is the better tradeoff -- I think
the mjs import problem can/will eventually be fixed